# splinterd/api/static/openapi.yaml
...
openapi: '3.0.0'
info:
  version: 0.5.1
  title: splinterd API
  description: REST API for the Splinter daemon
servers:
  - url: http://localhost:9000/api
paths:
...
  /ws/admin/register/{type}:
    get:
      tags:
        - Admin Service
      description: Register the handler for a circuit management type
      parameters:
        - $ref: "#/components/parameters/auth"
        - $ref: "#/components/parameters/protocol_version"
        - name: type
          description: The circuit management type is the type of circuit the handler will manage
          in: path
          required: true
          schema:
            type: string
          x-example: gameroom
        - name: last
          description: A timestamp in milliseconds from the Unix Epoch, indicating the last received event.
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Registration request was successfully submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationRegistration'
        400:
          description: |
            The circuit management type was invalid or the request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: The client is unauthorized
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    auth:
      name: Authorization
      in: header
      description: |
        The client's authorization, which the server resolves to an identity.
        Currently supports Biome JWT (if Biome credentials is enabled), Cylinder
        JWT, and OAuth2 bearer tokens.
      required: true
      schema:
        type: string
        examples:
          Biome: Bearer Biome:eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lk\
            IjoiZjM1YWFjYzEtYTljZC00ZWRhLWI2ZDAtMmVmYWRkZjBjOGE0IiwiaXNzIjoic2V\
            sZi1pc3N1ZWQiLCJleHAiOjE1ODAyMzkyMjh9.P8hA0ru_xriYX7qryl08ZEp86t5HD\
            _AEVPEUXY70Ehc
          Cylinder: Bearer Cylinder:eyJhbGciOiJzZWNwMjU2azEiLCJ0eXAiOiJjeWxpbmR\
            lcitqd3QifQ==.eyJpc3MiOiIwMjc5YmU2NjdlZjlkY2JiYWM1NWEwNjI5NWNlODcwY\
            jA3MDI5YmZjZGIyZGNlMjhkOTU5ZjI4MTViMTZmODE3OTgifQ==.71Vw3+m9R4b8iZU\
            zhRHOAtX/hO5WYA9PgMe27/CeZ6NhIXFYkBBzreoIGpHbfJ8UxT+1MLUgjsQB8TISaf\
            neRA==
          OAuth2: Bearer OAuth2:55eeea7ce2b472d69d406990939baa698e34e955
    protocol_version:
      name: SplinterProtocolVersion
      in: header
      description: |
        The protocol version which the client can understand. If not provided,
        the node will respond using its latest protocol version.
      required: false
      schema:
        type: integer
        example: 2
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    ApplicationRegistration:
      additionalProperties: false
      properties:
        status:
          description: Registration status
          type: string
          enum:
            - OK
            - ERROR
        circuit_management_type:
          description: Circuit management type
          type: string
          example: gameroom
        error_reason:
          description: Error message if the registration failed
          type: string
...